+++ src/infer/inferOs.ts
@@ -1,8 +1,9 @@
+import type { SupportedArch, SupportedPlatform } from '@electron/packager';
 import * as os from 'os';
 
 import * as log from 'loglevel';
 
-// Ideally we'd get this list directly from electron-packager, but it's not
+// Ideally we'd get this list directly from @electron/packager, but it's not
 // possible to convert a literal type to an array of strings in current TypeScript
 export const supportedArchs = ['x64', 'armv7l', 'arm64', 'universal'];
 export const supportedPlatforms = [
@@ -15,21 +16,21 @@ export const supportedPlatforms = [
   'windows',
 ];
 
-export function inferPlatform(): string {
+export function inferPlatform(): SupportedPlatform {
   const platform = os.platform();
   if (['darwin', 'linux', 'win32'].includes(platform)) {
     log.debug('Inferred platform', platform);
-    return platform;
+    return platform as SupportedPlatform;
   }
 
   throw new Error(`Untested platform ${platform} detected`);
 }
 
-export function inferArch(): string {
+export function inferArch(): SupportedArch {
   const arch = os.arch();
   if (!supportedArchs.includes(arch)) {
     throw new Error(`Incompatible architecture ${arch} detected`);
   }
   log.debug('Inferred arch', arch);
-  return arch;
+  return arch as SupportedArch;
 }